tweak(phone/Snackbar): move to os dir and tweak close handling
