+++ phone/src/apps/camera/components/grid/GalleryGrid.tsx
@@ -11,7 +11,7 @@ import { ServerPromiseResp } from '../../../../../../typings/common';
 import { GalleryPhoto, PhotoEvents } from '../../../../../../typings/photo';
 import { usePhotoActions } from '../../hooks/usePhotoActions';
 import { usePhotosValue } from '../../hooks/state';
-import { useSnackbar } from '../../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../../os/snackbar/hooks/useSnackbar';
 import { useTranslation } from 'react-i18next';
 
 export const GalleryGrid = () => {
