+++ phone/src/apps/camera/hooks/useCamera.ts
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { SetterOrUpdater, useRecoilState } from 'recoil';
 import { GalleryPhoto, PhotoEvents } from '../../../../../typings/photo';
 import { useNuiEvent, useNuiCallback } from 'fivem-nui-react-lib/';
-import { useSnackbar } from '../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../os/snackbar/hooks/useSnackbar';
 import { photoState } from './state';
 
 interface IUseCamera {
