+++ phone/src/apps/contacts/hooks/useContactsAPI.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useSnackbar } from '../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../os/snackbar/hooks/useSnackbar';
 import { fetchNui } from '../../../utils/fetchNui';
 import { ServerPromiseResp } from '../../../../../typings/common';
 import { Contact, ContactEvents, PreDBContact } from '../../../../../typings/contact';
