+++ phone/src/apps/marketplace/components/MarketplaceList/ListingActions.tsx
@@ -10,7 +10,7 @@ import PhoneIcon from '@mui/icons-material/Phone';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router';
 import { fetchNui } from '../../../../utils/fetchNui';
-import { useSnackbar } from '../../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../../os/snackbar/hooks/useSnackbar';
 import { ServerPromiseResp } from '../../../../../../typings/common';
 import { useMyPhoneNumber } from '../../../../os/simcard/hooks/useMyPhoneNumber';
 import { useCall } from '../../../../os/call/hooks/useCall';
