+++ phone/src/apps/messages/components/modal/MessageModal.tsx
@@ -23,7 +23,7 @@ import { Message, MessageEvents } from '../../../../../../typings/messages';
 import Modal from '../../../../ui/components/Modal';
 import { fetchNui } from '../../../../utils/fetchNui';
 import { ServerPromiseResp } from '../../../../../../typings/common';
-import { useSnackbar } from '../../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../../os/snackbar/hooks/useSnackbar';
 import { useMessageActions } from '../../hooks/useMessageActions';
 import { useMessagesState } from '../../hooks/state';
