+++ phone/src/apps/notes/hooks/useNotesAPI.ts
@@ -2,7 +2,7 @@ import { useCallback } from 'react';
 import { fetchNui } from '../../../utils/fetchNui';
 import { ServerPromiseResp } from '../../../../../typings/common';
 import { BeforeDBNote, DeleteNoteDTO, NoteItem, NotesEvents } from '../../../../../typings/notes';
-import { useSnackbar } from '../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../os/snackbar/hooks/useSnackbar';
 import { useTranslation } from 'react-i18next';
 import { useNotesActions } from './useNotesActions';
