+++ phone/src/apps/settings/components/WallpaperModal.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useCustomWallpaperModal, useSettings } from '../hooks/useSettings';
 import { useTranslation } from 'react-i18next';
 import DialogForm from '../../../ui/components/DialogForm';
-import { useSnackbar } from '../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../os/snackbar/hooks/useSnackbar';
 import { TextField } from '../../../ui/components/Input';
 
 const WallpaperModal: React.FC = () => {
