+++ phone/src/apps/twitter/components/AddTweetModal.tsx
@@ -16,9 +16,9 @@ import { getNewLineCount } from '../utils/message';
 import { NewTweet, TwitterEvents } from '../../../../../typings/twitter';
 import { fetchNui } from '../../../utils/fetchNui';
 import { ServerPromiseResp } from '../../../../../typings/common';
-import { useSnackbar } from '../../../ui/hooks/useSnackbar';
 import { useTranslation } from 'react-i18next';
 import { promiseTimeout } from '../../../utils/promiseTimeout';
+import { useSnackbar } from '../../../os/snackbar/hooks/useSnackbar';
 
 const useStyles = makeStyles((theme) => ({
   root: {
