+++ phone/src/apps/twitter/components/profile/Profile.tsx
@@ -9,7 +9,7 @@ import { TwitterEvents } from '../../../../../../typings/twitter';
 import ProfileField from '../../../../ui/components/ProfileField';
 import { fetchNui } from '../../../../utils/fetchNui';
 import { ServerPromiseResp } from '../../../../../../typings/common';
-import { useSnackbar } from '../../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../../../os/snackbar/hooks/useSnackbar';
 
 const useStyles = makeStyles(() => ({
   root: {
