+++ phone/src/apps/twitter/components/tweet/ShowMore.tsx
@@ -7,8 +7,8 @@ import ReportButton from '../buttons/ReportButton';
 import { TwitterEvents } from '../../../../../../typings/twitter';
 import { fetchNui } from '../../../../utils/fetchNui';
 import { ServerPromiseResp } from '../../../../../../typings/common';
-import { useSnackbar } from '../../../../ui/hooks/useSnackbar';
 import { useTwitterActions } from '../../hooks/useTwitterActions';
+import { useSnackbar } from '../../../../os/snackbar/hooks/useSnackbar';
 
 export const ShowMore = ({ id, isReported, isMine }) => {
   const { t } = useTranslation();
