+++ phone/src/os/call/hooks/useCall.ts
@@ -4,7 +4,7 @@ import { CallEvents } from '../../../../../typings/call';
 import { fetchNui } from '../../../utils/fetchNui';
 import { useCallback } from 'react';
 import { useMyPhoneNumber } from '../../simcard/hooks/useMyPhoneNumber';
-import { useSnackbar } from '../../../ui/hooks/useSnackbar';
+import { useSnackbar } from '../../snackbar/hooks/useSnackbar';
 import { useTranslation } from 'react-i18next';
 import { ServerPromiseResp } from '../../../../../typings/common';
