+++ typings/phone.ts
@@ -5,6 +5,7 @@ export enum PhoneEvents {
   OPEN_PHONE = 'npwd:open',
   CLOSE_PHONE = 'npwd:close',
   SET_VISIBILITY = 'npwd:setVisibility',
+  ADD_SNACKBAR_ALERT = 'npwd:setSnackarAlert',
   SET_NUMBER = 'npwd:setNumber',
   SET_PHONE_READY = 'npwd:phoneReady',
   SET_CONFIG = 'npwd:setPhoneConfig',
